# Create an 80.4% speedup when using borrowed str refs