--- a/src/lib/records.rs
+++ b/src/lib/records.rs
@@ -45,24 +45,24 @@ pub fn tumor_only_header(sample: String) -> Header {
 
 /// A record of output from VarDict/VarDictJava run in tumor-only mode.
 #[derive(Debug, Deserialize)]
-pub struct TumorOnlyVariant {
-    sample: String,
-    interval_name: String,
-    contig: String,
+pub struct TumorOnlyVariant<'a> {
+    sample: &'a str,
+    interval_name: &'a str,
+    contig: &'a str,
     start: u64,
     end: u64,
-    ref_allele: String,
-    alt_allele: String,
+    ref_allele: &'a str,
+    alt_allele: &'a str,
     depth: u32,
     alt_depth: u32,
     ref_forward: u32,
     ref_reverse: u32,
     alt_forward: u32,
     alt_reverse: u32,
-    gt: String,
+    gt: &'a str,
     af: f32,
     // Semi-colon separated integers, could be a struct
-    strand_bias: String,
+    strand_bias: &'a str,
     mean_position_in_read: f32,
     stdev_position_in_read: f32,
     mean_base_quality: f32,
@@ -80,20 +80,20 @@ pub struct TumorOnlyVariant {
     mean_mismatches_in_reads: f32,
     high_quality_variant_reads: u32,
     high_quality_total_reads: u32,
-    flank_seq_5_prime: String,
-    flank_seq_3_prime: String,
+    flank_seq_5_prime: &'a str,
+    flank_seq_3_prime: &'a str,
     // Position format, unused in VCF.
-    segment: String,
-    variant_type: String,
-    duplication_rate: String,
+    segment: &'a str,
+    variant_type: &'a str,
+    duplication_rate: &'a str,
     // Either a zero, or triple of ints separated with "-".
     // We can deserialize this further into a struct of SV details.
-    sv_details: String,
+    sv_details: &'a str,
     #[serde(default)]
-    distance_to_crispr_site: Option<String>,
+    distance_to_crispr_site: Option<u32>,
 }
 
-impl AbstractInterval for TumorOnlyVariant {
+impl<'a> AbstractInterval for TumorOnlyVariant<'a> {
     fn contig(&self) -> &str {
         &self.contig
     }
@@ -104,23 +104,23 @@ impl AbstractInterval for TumorOnlyVariant {
 
 /// A record of output from VarDict/VarDictJava run in amplicon-aware mode.
 #[derive(Debug, Deserialize)]
-struct AmpliconVariant {
-    sample: String,
-    interval_name: String,
-    contig: String,
+struct AmpliconVariant<'a> {
+    sample: &'a str,
+    interval_name: &'a str,
+    contig: &'a str,
     start: u64,
     end: u64,
-    ref_allele: String,
-    alt_allele: String,
+    ref_allele: &'a str,
+    alt_allele: &'a str,
     depth: u32,
     alt_depth: u32,
     ref_forward: u32,
     ref_reverse: u32,
     alt_forward: u32,
     alt_reverse: u32,
-    gt: String,
+    gt: &'a str,
     af: f32,
-    strand_bias: String,
+    strand_bias: &'a str,
     mean_position_in_read: f32,
     stdev_position_in_read: f32,
     mean_base_quality: f32,
@@ -138,11 +138,11 @@ struct AmpliconVariant {
     mean_mismatches_in_reads: f32,
     high_quality_variant_reads: u32,
     high_quality_total_reads: u32,
-    flank_seq_5_prime: String,
-    flank_seq_3_prime: String,
+    flank_seq_5_prime: &'a str,
+    flank_seq_3_prime: &'a str,
     // Position format, unused in VCF.
-    segment: String,
-    variant_type: String,
+    segment: &'a str,
+    variant_type: &'a str,
     num_amplicons_supporting_variant: u32,
     total_amplicons_overlapping: u32,
     num_amplicons_rare: u32,
@@ -150,7 +150,7 @@ struct AmpliconVariant {
     top_variant_in_amplicon_does_not_match: bool,
 }
 
-impl AbstractInterval for AmpliconVariant {
+impl<'a> AbstractInterval for AmpliconVariant<'a> {
     fn contig(&self) -> &str {
         &self.contig
     }