The command:

```bash
❯ vartovcf -i test.var | wc -l
[2021-01-26T01:47:32Z INFO  vartovcflib::vartovcf] Input file:  "test.var"
[2021-01-26T01:47:32Z INFO  vartovcflib::vartovcf] Output file: "/dev/stdout"
[2021-01-26T01:47:33Z INFO  vartovcflib::vartovcf] Processed 22353 variant records
22353
```

Release binary and previous commit:

```bash
❯ hyperfine --warmup 5 'vartovcf -i test.var'
Benchmark 1: vartovcf -i test.var
  Time (mean ± σ):      1.166 s ±  0.034 s    [User: 1.141 s, System: 0.023 s]
  Range (min … max):    1.109 s …  1.231 s    10 runs
```

Release binary and after amortization of string data:

```bash
❯ hyperfine --warmup 5 'vartovcf -i test.var'
Benchmark 1: vartovcf -i test.var
  Time (mean ± σ):     933.4 ms ±  25.6 ms    [User: 912.3 ms, System: 19.0 ms]
  Range (min … max):   901.7 ms … 966.8 ms    10 runs
```